Restructured the examples folder

+++ examples/datasets/get_data.py
@@ -4,7 +4,7 @@ import os
 folder_path = os.path.dirname(os.path.realpath(__file__))
 print('Beginning download of datasets')
 
-datasets = ['AllNLI.zip', 'stsbenchmark.zip', 'wikipedia-sections-triplets.zip']
+datasets = ['AllNLI.zip', 'stsbenchmark.zip', 'wikipedia-sections-triplets.zip', 'STS2017.en-de.txt.gz', 'TED2013-en-de.txt.gz', 'xnli-en-de.txt.gz']
 server = "https://public.ukp.informatik.tu-darmstadt.de/reimers/sentence-transformers/datasets/"
 
 for dataset in datasets:
@@ -13,10 +13,11 @@ for dataset in datasets:
     dataset_path = os.path.join(folder_path, dataset)
     urllib.request.urlretrieve(url, dataset_path)
 
-    print("Extract", dataset)
-    with zipfile.ZipFile(dataset_path, "r") as zip_ref:
-        zip_ref.extractall(folder_path)
-    os.remove(dataset_path)
+    if dataset.endswith('.zip'):
+        print("Extract", dataset)
+        with zipfile.ZipFile(dataset_path, "r") as zip_ref:
+            zip_ref.extractall(folder_path)
+        os.remove(dataset_path)
 
 
 print("All datasets downloaded and extracted")
